requirements.txt
PyOpenGL==3.0.2
# pyobjc==2.2 # OSX only, required by pyglet
pyglet==1.2.4
